Restructure generators
