+++ webpack/generator/templates/component/component.test.js.hbs
@@ -1,4 +1,4 @@
-import {{ properCase name }} from '../index';
+import {{ properCase name }} from './index';
 
 import expect from 'expect';
 import { shallow } from 'enzyme';
