+++ webpack/generator/templates/component/es6.js.hbs
