+++ webpack/generator/templates/component/stateless.js.hbs
