+++ webpack/generator/templates/component/styles.css.hbs
