+++ webpack/generator/templates/container/actions.js.hbs
