+++ webpack/generator/templates/container/actions.test.js.hbs
