+++ webpack/generator/templates/container/constants.js.hbs
