+++ webpack/generator/templates/container/container.test.js.hbs
