+++ webpack/generator/templates/container/index.js.hbs
