+++ webpack/generator/templates/container/reducer.js.hbs
