+++ webpack/generator/templates/container/reducer.test.js.hbs
@@ -1,6 +1,5 @@
 import expect from 'expect';
 import {{ properCase name }}Reducer from '../reducer';
-import {  } from '../constants';
 
 describe('{{ properCase name }}Reducer', () => {
   it('returns the initial state', () => {
