+++ webpack/generator/templates/container/selector.js.hbs
