+++ webpack/generator/templates/container/styles.css.hbs
