+++ webpack/generator/templates/route.hbs
