+++ webpack/generator/templates/saga.js.hbs
