+++ webpack/generator/templates/saga.test.js.hbs
