+++ webpack/generator/templates/selector.js.hbs
