+++ webpack/generator/templates/selector.test.js.hbs
